Issue#3: Complete Database First Draft
This is the completion of the database first draft. The result is an
accurate database model in MySQL syntax in database_design.txt, and
create_db.py and drop_db.py scripts in a db_scripts folder. These are
just prototype files - the production DB will probably be sqlite in
flask-sqlalchemy.